@@ -15,54 +15,65 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.1.0",
-    "@angular/animations": "^9.0.5",
-    "@angular/cdk": "^9.1.3",
-    "@angular/common": "~9.0.5",
-    "@angular/compiler": "~9.0.5",
-    "@angular/core": "~9.0.5",
+    "@angular/animations": "^9.1.0",
+    "@angular/cdk": "^9.2.0",
+    "@angular/common": "^9.1.0",
+    "@angular/compiler": "~9.1.0",
+    "@angular/core": "^9.1.0",
     "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "~9.0.5",
-    "@angular/material": "^9.1.2",
-    "@angular/platform-browser": "~9.0.5",
-    "@angular/platform-browser-dynamic": "~9.0.5",
-    "@angular/router": "~9.0.5",
-    "@fortawesome/fontawesome-free": "^5.12.1",
+    "@angular/forms": "~9.1.0",
+    "@angular/material": "^9.2.0",
+    "@angular/platform-browser": "~9.1.0",
+    "@angular/platform-browser-dynamic": "~9.1.0",
+    "@angular/router": "~9.1.0",
+    "@fortawesome/fontawesome-free": "^5.13.0",
+    "@popperjs/core": "^2.1.1",
     "@types/chart.js": "^2.9.16",
     "@types/googlemaps": "^3.39.3",
     "angular-bootstrap-md": "^9.0.0",
     "animate.css": "^3.7.2",
+    "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
-    "chart.js": "^2.5.0",
+    "bootstrap": "^4.4.1",
+    "chart.js": "^2.9.3",
     "concurrently": "^5.1.0",
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
+    "jquery": "^3.4.1",
+    "jsonwebtoken": "^8.5.1",
     "mongodb": "^3.5.4",
-    "mongoose": "^5.9.3",
+    "mongoose": "^5.9.7",
+    "mongoose-unique-validator": "^2.0.3",
+    "ngx-page-scroll": "^7.0.1",
+    "ngx-page-scroll-core": "^7.0.1",
+    "npm": "^6.14.4",
+    "popper.js": "^1.16.1",
     "rxjs": "~6.5.4",
     "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.5",
-    "@angular/cli": "~9.0.5",
-    "@angular/compiler-cli": "~9.0.5",
-    "@angular/language-service": "~9.0.5",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/build-angular": "~0.901.0",
+    "@angular/cli": "~9.1.0",
+    "@angular/compiler-cli": "~9.1.0",
+    "@angular/language-service": "~9.1.0",
+    "@types/jasmine": "^3.5.10",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
+    "@types/node": "^13.9.5",
+    "codelyzer": "^5.2.2",
     "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.3.0",
+    "jasmine-spec-reporter": "^5.0.1",
+    "karma": "^4.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-jasmine": "^3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.3",
     "nodemon": "^2.0.2",
     "protractor": "~5.4.3",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.7.5"
+    "ts-node": "^8.8.1",
+    "tslint": "^6.1.0",
+    "typescript": "^3.8.3"
   }
 }
