@@ -1,15 +1,24 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit } from "@angular/core";
+import { DataService } from "../../data.service";
+import { Game } from "../../game/game";
+import { HttpClient } from "@angular/common/http";
 
 @Component({
-  selector: 'app-page-games',
-  templateUrl: './page-games.component.html',
-  styleUrls: ['./page-games.component.scss']
+  selector: "app-page-games",
+  templateUrl: "./page-games.component.html",
+  styleUrls: ["./page-games.component.scss"]
 })
 export class PageGamesComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
+  games: Game[];
+  constructor(private _dataService: DataService, private http: HttpClient) {
+    // this._dataService.getGames().subscribe(res => (this.games = res));
   }
 
+  // ngOnInit(): void {}
+
+  ngOnInit() {
+    this.http
+      .get("assets/game-object.json")
+      .subscribe((data: []) => (this.games = data));
+  }
 }
